+++ src/db/mod.rs
@@ -14,6 +14,7 @@ pub mod imap_state;
 pub mod init;
 pub mod maildir_ids;
 pub mod managesieve_ids;
+pub mod roles;
 pub mod sources;
 pub mod sync_state_jmap;
 pub mod takeout_ids;