+++ CHANGELOG.md
@@ -2,6 +2,19 @@
 
 All notable changes to this project will be documented in this file. This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.0.3] - 2026-06-XX
+
+### Added
+
+### Changed
+
+### Fixed
+- Mailbox roles must be unique per archive (#8).
+
+# Change Log
+
+All notable changes to this project will be documented in this file. This project adheres to [Semantic Versioning](http://semver.org/).
+
 ## [1.0.2] - 2026-06-11
 
 ### Added